requirements.txt
Django
psycopg2-binary
requests
pytesseract
Pillow
tesseract
tesseract-ocr